Update game.py
if the user by mistake uses the same block for example
X | X | O
- + - + -
  | O | X
- + - + -
O |   |  
user input is 2 where it is already used and it is wrong input in such case if the user repeats the same mistake then there will be the only total of 9 inputs and the game is incomplete and will be stopping in middle. so this is the fixing of the bug. in this there will be a total of 9 correct inputes if wrong not counted.